+++ tests/unit/protocol-menu.js
@@ -88,14 +88,14 @@ describe('protocol-menu.js', function() {
             item.dispatchEvent(mockEnterEvent);
             jasmine.clock().tick(200);
 
-            expect(item.classList.contains('is-selected')).toBeTruthy();
+            expect(item.classList.contains('mzp-is-selected')).toBeTruthy();
             expect(options.open).toHaveBeenCalled();
 
             mockLeaveEvent.initEvent('mouseleave', true, true);
             item.dispatchEvent(mockLeaveEvent);
             jasmine.clock().tick(200);
 
-            expect(item.classList.contains('is-selected')).toBeFalsy();
+            expect(item.classList.contains('mzp-is-selected')).toBeFalsy();
             expect(options.close).toHaveBeenCalled();
         });
 
@@ -113,7 +113,7 @@ describe('protocol-menu.js', function() {
 
             title.click();
 
-            expect(item.classList.contains('is-selected')).toBeTruthy();
+            expect(item.classList.contains('mzp-is-selected')).toBeTruthy();
             expect(options.open).toHaveBeenCalled();
         });
 
@@ -134,12 +134,12 @@ describe('protocol-menu.js', function() {
             item.dispatchEvent(mockEnterEvent);
             jasmine.clock().tick(200);
 
-            expect(item.classList.contains('is-selected')).toBeTruthy();
+            expect(item.classList.contains('mzp-is-selected')).toBeTruthy();
             expect(options.open).toHaveBeenCalled();
 
             document.querySelector('.mzp-c-menu-button-close').click();
 
-            expect(item.classList.contains('is-selected')).toBeFalsy();
+            expect(item.classList.contains('mzp-is-selected')).toBeFalsy();
             expect(options.buttonClose).toHaveBeenCalled();
             expect(options.close).toHaveBeenCalled();
         });
@@ -172,11 +172,11 @@ describe('protocol-menu.js', function() {
             });
 
             title.click();
-            expect(item.classList.contains('is-selected')).toBeTruthy();
+            expect(item.classList.contains('mzp-is-selected')).toBeTruthy();
             expect(options.open).toHaveBeenCalled();
 
             title.click();
-            expect(item.classList.contains('is-selected')).toBeFalsy();
+            expect(item.classList.contains('mzp-is-selected')).toBeFalsy();
             expect(options.close).toHaveBeenCalled();
         });
     });